Replaces one HTML tag with another
When creating HTML emails, you find quickly that email clients are way
behind web browsers when it comes to rendering HTML, and on HTML
standards.

One issue is default spacing below certain types of block elements in
certain email clients. When testing an HTML email, and unusual spacing
is observed, changing to different element types can be a simple
solution.

This tool was created as a simple tag replacer, using PHP string
replace functions. It’s quick and easy to use on snippets of or HTML
code.